Fix bugs in note errors checker
Handle V1 timings declared in seconds (float) instead of ms (int)
Regenerate DB migrations according to new typing
Fix EOP tag parsing when sometimes endbeat/mutliple beats are specified (just ignore them)
Make LibrariesController returning all songs by default
Add last parsed time property on song
Add the current line in alerts when available
Add overlap notes detection

+++ Back/KaraWeb.Shared/Models/Songs/SongDtoBase.cs
@@ -41,19 +41,19 @@ namespace KaraWeb.Shared.Models.Songs
         ///     The GAP between start of the audio and first beat
         /// </summary>
         /// <remarks>milliseconds</remarks>
-        public int? Gap { get; set; }
+        public double? Gap { get; set; }
 
         /// <summary>
         ///     The GAP between start of the audio and first beat
         /// </summary>
         /// <remarks>milliseconds</remarks>
-        public int? Start { get; set; }
+        public double? Start { get; set; }
 
         /// <summary>
         ///     The GAP between start of the audio and first beat
         /// </summary>
         /// <remarks>milliseconds</remarks>
-        public int? End { get; set; }
+        public double? End { get; set; }
 
         /// <summary>
         ///     Song's players defined from #P1 to #P9
@@ -73,13 +73,13 @@ namespace KaraWeb.Shared.Models.Songs
         ///     The video delay relative to the audio file
         /// </summary>
         /// <remarks>milliseconds</remarks>
-        public int? VideoGap { get; set; }
+        public double? VideoGap { get; set; }
 
         /// <summary>
         ///     The offset of audio file to use for preview
         /// </summary>
         /// <remarks>milliseconds</remarks>
-        public int? PreviewStart { get; set; }
+        public double? PreviewStart { get; set; }
 
         /// <summary>
         ///     The offset of audio file to use to start in a medley
